fix(security): prevent JWT expiry int overflow and enforce isEnabled on JWT auth filter
- JwtUtils: change jwtExpirationMs from int to long; int overflows silently at
  ~24.8 days causing tokens to expire immediately or carry a negative expiry date
- AuthTokenFilter: skip SecurityContext population when UserDetails.isEnabled()
  returns false; previously a deactivated account (active=false) remained
  authenticated for the full JWT lifetime (up to 24h) because the filter never
  checked the flag — only the login flow did

+++ src/main/java/br/com/cameraeluz/acervo/security/JwtUtils.java
@@ -23,7 +23,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${photoarchive.app.jwtExpirationMs}")
-    private int jwtExpirationMs;
+    private long jwtExpirationMs;
 
     /**
      * Generates a JWT token based on the authenticated user details.
